Reapply the diff between the jvstm-fenix and the plain jvstm to clean git history
This commit just redoes the changes made a long time ago to the jvstm-fenix, namely, to
use an HashMap instead of a Cons to keep the readset.  The purpose of this commit is
to clean the history of git.  From now on, all changes to the jvstm should be made on
the jvstm-master branch and merged back into the jvstm-fenix branch.  Making changes
directly on the jvstm-fenix branch makes sense only for things that should apply only
to the fenix version.