Introduce a new Callable-based ProcessAtomicAnnotations
The default behavior of the new code is unchanged from the old ProcessAtomicAnnotations.
JVSTM clients do not need any modifications to continue working.

For each method tagged with @Atomic, a new static final field is inserted into its class, which
contains an AtomicContext instance (the default one mimicking the old PAA behavior).
The tagged method itself does not contain the transaction begin/commit/abort loop inlined anymore,
instead instantiating a new Callable, and invoking the doTransactionally method of the context,
passing it the created Callable.
This Callable, when invoked, calls the real original method code, which was preserved in a new
method with its name prefixed with atomic$.

The new structure allows easy modifications to the behavior of the @Atomic annotation:
- The default @Atomic behavior, previously hardcoded on PAA (and exemplified in AtomicSample.java),
is now specified in Java (on the DefaultAtomicContext class)
- Each @Atomic method can have a different behavior, selectable as an argument of the @Atomic
annotation (future expansion may allow runtime replacement of the behavior)
- New behaviors can be added by implementing the ContextFactory and AtomicContext interfaces

Java 7 is now fully supported.

Some implementation notes:
- ProcessAtomicAnnotations and companion classes have been moved into the (new) jvstm.atomic package
- Upgraded the ASM libraries to version 4
- When a ContextFactory is being asked for a new AtomicContext instance, it receives an instance of
the @Atomic annotation, so that any parameters given there are available to the ContextFactory.
Due to some uninteresting details, this instance is automatically generated, and not available as
Java source. This generation is done automatically when creating the JVSTM jar, by the
GenerateAtomicInstance class --- any added fields to the @Atomic annotation are automatically
picked up and supported by the generated class.
- Each Callable is also automatically generated, and placed on the same package as the class
containing the original method
- For more details "Use the source, Luke"

Work done in collaboration with Hugo Rito <hugo.rito@ist.utl.pt>