Fix infinite loop due to bad interaction between the early abort feature and nested transactions
When the early abort feature was added, it used a CommitException to signal that a transaction
should abort and restart. This "overloaded" the usual CommitException semantics, where a
CommitException was used to signal that the commit operation of a Transaction had failed.

This could lead to an infinite loop with the default behavior of @Atomic annotation, or any code
with similar structure, when an early abort occurred from inside a nested transaction.

Also tweaked the constructors of JVSTM exceptions to be protected as they should not be part of the
public API.

The following testcase would never finish with the old code, in a JVSTM version with early aborts:

public class Test extends Thread {
    static jvstm.VBoxInt counter = new jvstm.VBoxInt(0);

    public static void main(String[] args) {
        new Test(1000).start();
        new Test(0).start();
    }

    int sleep;
    Test(int s) { sleep = s; }

    @jvstm.Atomic(speculativeReadOnly=false)
    public void run() {
        try { Thread.sleep(sleep); } catch (Exception e) { }
        doNested();
    }

    @jvstm.Atomic
    void doNested() { counter.inc(); }
}

Work done in collaboration with Hugo Rito <hugo.rito@ist.utl.pt>