Fix Unit Tests jvstm.test.point.utests.gc
These unit tests check the behavior of the reversion process (converting an object from the extended layout---i.e. with a versions’ history---to the compact layout---i.e. with the field VBox::body pointing to null). The process of reverting an object is plugged into the JVSTM’s algorithm for garbage collecting (GC) old values from the objects’ histories. So, once the GC decides to trim an object’s history, it may happen that only a single version remains (the most recent, of course), in which case we may revert the object to the compact layout.

Furthermore, the JVSTM’s GC algoritm ensures that a version in the history of an object is not removed as long as there may be some running transaction that may need to access it. So, if for some reason there is a pending transaction from an unfinished previous test, then the GC will not run in the current tests for the jvstm.test.point.utests.gc and therefore the reversion will be not dispatched, invalidating the expected results.

* AomGcTest (base class for these unit tests) => add a setup method that ensure that there is no pending transactions preventing the GC to run.

* TxContext => add a new next() property.