Update VBox's CASbody() to account for reloads
In the pure JVSTM the CAS can only fail because another thread already
committed this value.  However, when used together with Fenix Framework, it is
possible that the body changes because of reloads.  We identify this by
testing whether our commit did make it (this.body.version must be greater than
or equal to newValue.version).  If not, we retry the CAS.  This does not occur
in the pure JVSTM, which will just try the CAS once and even if it fails will
see the required body immediately.